repaired terminate process error
Not sure if totally fixed. The error was being caused by the processing if waiting requests after the simulation had ended during the after replication termination process.  This processing was check for and stopped. The base error appeared when the order of model element after replication calls caused the request queues not be cleared before releasing logic called during after replication termination processes.